Refs #25898. Move markers to the plotting module
