+++ qt/python/mantidqt/plotting/markers.py
